refactor: share the parameters of the input and output sockets of varying
